ingestion/requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
minio==7.2.7
pyarrow==17.0.0
duckdb==1.0.0
